move update_effect_config to virtuals.py
